@@ -15,26 +15,26 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "babel-plugin-styled-components": "^1.1.7",
-    "eslint-plugin-import": "^2.7.0",
-    "express": "^4.15.2",
-    "next": "beta",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
-    "styled-components": "^2.1.2",
+    "babel-plugin-styled-components": "^1.2.1",
+    "eslint-plugin-import": "^2.8.0",
+    "express": "^4.16.2",
+    "next": "^4.1.3",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
+    "styled-components": "^2.2.1",
     "unfetch": "^3.0.0",
-    "webpack": "^2.5.1"
+    "webpack": "^2.7.0"
   },
   "devDependencies": {
-    "eslint": "^4.4.0",
+    "eslint": "^4.9.0",
     "eslint-config-airbnb": "^15.1.0",
-    "eslint-config-prettier": "^2.3.0",
+    "eslint-config-prettier": "^2.6.0",
     "eslint-plugin-jsx-a11y": "^6.0.2",
-    "eslint-plugin-prettier": "^2.1.2",
-    "eslint-plugin-react": "^7.1.0",
-    "prettier": "^1.5.3",
-    "stylelint": "^8.0.0",
+    "eslint-plugin-prettier": "^2.3.1",
+    "eslint-plugin-react": "^7.4.0",
+    "prettier": "^1.7.4",
+    "stylelint": "^8.2.0",
     "stylelint-config-standard": "^17.0.0",
-    "stylelint-processor-styled-components": "^0.2.1"
+    "stylelint-processor-styled-components": "^0.2.2"
   }
 }
