@@ -1,7 +1,6 @@
 import convertUnit from "../css-convert-unit.js";
 import pxToRem from "../px-to-rem.js";
-
-const variables = require("../css-variables.js");
+import { TYPEUNITS } from "../css-variables.js";
 
 /* TODO: Allow this to accept a spacing scale, not just type scale.
  * As of now, it only looks for type size and scales.
@@ -9,32 +8,44 @@ const variables = require("../css-variables.js");
 
 /* PARAMETERS
 
- * type:  boolean (e.g. type  || spacing)
- * size: boolean (e.g. small || large)
- * direction:  integer (e.g. 1, -2, 4)                                        */
-export default function getModularScale(direction, scale) {
+ * typeOfScale:      boolean (typography || spacing)
+ * sizeOfScale:      boolean (small || large)
+ * direction: integer                                                        */
+export default function getModularScale(typeOfScale, sizeOfScale, direction) {
+  if (!typeOfScale) typeOfScale = "typography";
+  if (!sizeOfScale) sizeOfScale = "small";
+
   let modularSize = modularSize;
+  let baseSize = baseSize;
+  let scale = scale;
 
-  let baseSize = variables.TYPEUNITS.bodyFont.small;
-  let typescale = variables.TYPEUNITS.typographyScale.small;
-
-  if (scale === 'large') {
-    baseSize = variables.TYPEUNITS.bodyFont.large;
-    typescale = variables.TYPEUNITS.typographyScale.large;
+  if (typeOfScale === "typography") {
+    if (sizeOfScale === "small") {
+      baseSize = TYPEUNITS.bodyFont.small;
+      scale = TYPEUNITS.typographyScale.small;
+    } else {
+      baseSize = TYPEUNITS.bodyFont.large;
+      scale = TYPEUNITS.typographyScale.large;
+    }
+  } else if (typeOfScale === "spacing") {
+    baseSize = TYPEUNITS.bodyFont.lineHeight;
+    if (sizeOfScale === "small") {
+      scale = TYPEUNITS.spacingScale.small;
+    } else {
+      scale = TYPEUNITS.spacingScale.large;
+    }
   }
 
   if (direction < 0) {
-    modularSize = baseSize / typescale * direction * -1;
+    modularSize = baseSize / scale * direction * -1;
   } else if (direction === 0) {
     modularSize = baseSize;
   } else if (direction > 0) {
-    modularSize = Math.pow(typescale, direction) * baseSize;
+    modularSize = Math.pow(scale, direction) * baseSize;
   }
 
-
-  const valueInPX = Math.floor(modularSize);
+  const valueInPX = Math.round(modularSize);
   const valueInREM = pxToRem(valueInPX);
 
   return valueInREM;
-
 }
